+++ src/glamp_owners/models.py
@@ -11,7 +11,7 @@ User = get_user_model()
 
 
 class GlampOwner(TimestampMixin):
-    user = models.OneToOneField(User, on_delete=models.CASCADE)
+    user = models.OneToOneField(User, on_delete=models.CASCADE, default=None, null=True)
 
     first_name = models.CharField(max_length=255, null=True, default=None)
     last_name = models.CharField(max_length=255, null=True, default=None)