+++ src/glamps/permissions.py
@@ -3,7 +3,7 @@ from django.contrib.auth.models import AnonymousUser
 from rest_framework.exceptions import PermissionDenied
 from rest_framework.permissions import BasePermission
 
-from glamps.models.glamp import Glamp
+from glamps.models import Glamp
 from roles.constants import Role
 
 User = get_user_model()